+++ CHANGELOG.md
@@ -23,6 +23,9 @@
     - Added addtional verbose logging.
 - Changes to SqlServerSecureConnection
   - Forced $Thumbprint to lowercase to fix [issue #1350](https://github.com/PowerShell/SqlServerDsc/issues/1350).
+  - Add parameter SuppressRestart with default value false.
+    This allows users to suppress restarts after changes have been made.
+    Changes will not take effect until the service has been restarted.
 - Changes to SqlSetup
   - Correct minor style violation [issue #1387](https://github.com/PowerShell/SqlServerDsc/issues/1387).
 - Changes to SqlDatabase